# Optimize `moveOp[Up,Down]InBlock` functions in `SimplifyGlobalAccesses`.